requirements.txt
openai
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
pandas
python-dotenv
tabulate
openpyxl
flask
gunicorn
sqlalchemy
psycopg2-binary
pgvector
pypdf2
